Refactor Dragger to 1 comp with 2 models.
