+++ desktop/cmp/panel/impl/dragger/Dragger.scss
