+++ desktop/cmp/panel/impl/dragger/DraggerModel.js
@@ -5,32 +5,12 @@
  * Copyright © 2019 Extremely Heavy Industries Inc.
  */
 
-import {hoistCmp, HoistModel, useLocalModel} from '@xh/hoist/core';
-import {div} from '@xh/hoist/cmp/layout';
+import {HoistModel} from '@xh/hoist/core';
 import {throwIf} from '@xh/hoist/utils/js';
 
-import './Dragger.scss';
-
-
-export const dragger = hoistCmp.factory({
-    displayName: 'Dragger',
-    model: false,
-
-    render({model}) {
-        const dragModel = useLocalModel(() => new DragModel(model));
-
-        return div({
-            className: `xh-resizable-dragger ${model.side}`,
-            onDrag: dragModel.onDrag,
-            onDragStart: dragModel.onDragStart,
-            onDragEnd: dragModel.onDragEnd,
-            draggable: true
-        });
-    }
-});
 
 @HoistModel
-class DragModel {
+export class DraggerModel {
 
     model;
     resizeState = null;
